+++ src/main/scala/aoc2019/day7/Amplifier.scala
@@ -4,13 +4,13 @@ import aoc2019.program.Program
 
 import scala.collection.mutable
 
-class Amplifier(program: Program, phase: Int) {
+class Amplifier(program: Program, phase: Long) {
 
-  private val inputs: mutable.Queue[Int] = mutable.Queue(phase)
+  private val inputs: mutable.Queue[Long] = mutable.Queue(phase)
 
   private val execution = program.start(() => inputs.dequeue())
 
-  def run(input: Int): Option[Int] = {
+  def run(input: Long): Option[Long] = {
     inputs.enqueue(input)
     execution.continue()
   }
